% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_search_volume}
\alias{get_search_volume}
\title{Get Search Volume Data for Keywords}
\usage{
get_search_volume(keywords, api_key, metrics_location = NULL,
  metrics_language = NULL, metrics_network = c("googlesearchnetwork",
  "googlesearch"), metrics_currency = NULL, method = c("get", "post"))
}
\arguments{
\item{keywords}{A vector of keywords to analyse}

\item{api_key}{API key from keywordtool.io}

\item{metrics_location}{Optional vector of location codes to get results from (see documentation for codes). Maximum 10 locations.}

\item{metrics_language}{Optional vector of language to search with. See `language_codes` for list of codes. Maximum 5 languages.}

\item{metrics_network}{Search network that will be used to pull the search volume data.}

\item{metrics_currency}{The currency that will be used to display cost-per-click (CPC) data. See 'currency_codes` for list of codes.}

\item{method}{Whether to request using get or post. Post is recommended for large requests.}
}
\value{
An object of class 'keyword_tool_result'.
}
\description{
Get search volume data though time, along with CPC and CMP data
for a set of up to 800 keywords.
}
\details{
For more on how to use this see the Keyword Tool API documentation:
http://keywordtool.io/api/documentation
}
\examples{
\dontrun{

keyword_tool_key <- '1234'

results <-
get_search_volume(keywords = c('scottish salmon', 'scotland fishing'),
                  api_key = keyword_tool_key,
                  metrics_location = c(20339,20342, 2840), #Scotland, England and US
                  metrics_network = 'googlesearch',
                  metrics_currency = 'GBP')

 str(results$content)

}
}

